Fix responsive table unit test imports, move helpers
