--- src/applications/disability-benefits/view-payments/tests/components/view-payments-lists/helpers.js
@@ -1,116 +1,0 @@
-export const payments = {
-  payments: [
-    {
-      payCheckAmount: '$3,261.10',
-      payCheckDt: '04/01/2019',
-      payCheckId: '001',
-      payCheckReturnFiche: 'C',
-      payCheckType: 'Compensation & Pension - Recurring',
-    },
-    {
-      payCheckAmount: '$3,261.10',
-      payCheckDt: '03/01/2019',
-      payCheckId: '002',
-      payCheckReturnFiche: 'C',
-      payCheckType: 'Compensation & Pension - Recurring',
-    },
-    {
-      payCheckAmount: '$177.60',
-      payCheckDt: '11/09/2018',
-      payCheckId: '003',
-      payCheckReturnFiche: 'C',
-      payCheckType: 'Compensation & Pension - Retroactive',
-    },
-    {
-      payCheckAmount: '$3,261.10',
-      payCheckDt: '04/01/2019',
-      payCheckId: '001',
-      payCheckReturnFiche: 'C',
-      payCheckType: 'Compensation & Pension - Recurring',
-    },
-    {
-      payCheckAmount: '$3,261.10',
-      payCheckDt: '03/01/2019',
-      payCheckId: '002',
-      payCheckReturnFiche: 'C',
-      payCheckType: 'Compensation & Pension - Recurring',
-    },
-    {
-      payCheckAmount: '$177.60',
-      payCheckDt: '11/09/2018',
-      payCheckId: '003',
-      payCheckReturnFiche: 'C',
-      payCheckType: 'Compensation & Pension - Retroactive',
-    },
-    {
-      payCheckAmount: '$3,261.10',
-      payCheckDt: '04/01/2019',
-      payCheckId: '001',
-      payCheckReturnFiche: 'C',
-      payCheckType: 'Compensation & Pension - Recurring',
-    },
-    {
-      payCheckAmount: '$3,261.10',
-      payCheckDt: '03/01/2019',
-      payCheckId: '002',
-      payCheckReturnFiche: 'C',
-      payCheckType: 'Compensation & Pension - Recurring',
-    },
-    {
-      payCheckAmount: '$177.60',
-      payCheckDt: '11/09/2018',
-      payCheckId: '003',
-      payCheckReturnFiche: 'C',
-      payCheckType: 'Compensation & Pension - Retroactive',
-    },
-  ],
-  returnPayments: [
-    {
-      returnedCheckAmount: null,
-      returnedCheckCancelDt: '12/01/2012',
-      returnedCheckIssueDt: null,
-      returnedCheckNum: '12365494',
-      returnedCheckRo: '17',
-      returnedCheckReason: '6',
-      returnedCheckReturnFiche: 'B',
-      returnedCheckType: 'Other',
-    },
-    {
-      returnedCheckAmount: '$50.00',
-      returnedCheckCancelDt: null,
-      returnedCheckIssueDt: '11/25/2012',
-      returnedCheckNum: '12365494',
-      returnedCheckRo: '17',
-      returnedCheckReason: '6',
-      returnedCheckReturnFiche: 'B',
-      returnedCheckType: 'Post-9/11 Educational Assistance',
-    },
-  ],
-};
-
-export const fields = [
-  {
-    label: 'Date',
-    value: 'date',
-  },
-  {
-    label: 'Amount',
-    value: 'amount',
-  },
-  {
-    label: 'Type',
-    value: 'type',
-  },
-  {
-    label: 'Method',
-    value: 'method',
-  },
-  {
-    label: 'Bank name',
-    value: 'bank',
-  },
-  {
-    label: 'Account',
-    value: 'account',
-  },
-];
