+++ src/applications/disability-benefits/view-payments/tests/helpers.js
@@ -1,146 +1,116 @@
-export const mockData = {
-  received: [
+export const payments = {
+  payments: [
     {
-      id: 1,
-      date: 'June 4, 2020',
-      amount: '$4,321',
-      type: 'Componsation & Pension Recurring',
-      method: 'Direct Deposit',
-      bank: 'USAA Federal Savings Bank',
-      account: '1234567890999',
-      rowClass: 'class',
+      payCheckAmount: '$3,261.10',
+      payCheckDt: '04/01/2019',
+      payCheckId: '001',
+      payCheckReturnFiche: 'C',
+      payCheckType: 'Compensation & Pension - Recurring',
     },
     {
-      id: 2,
-      date: 'June 3, 2020',
-      amount: '$4,321',
-      type: 'Componsation & Pension Recurring',
-      method: 'Direct Deposit',
-      bank: 'USAA Federal Savings Bank',
-      account: '1234567890999',
-      rowClass: 'class',
+      payCheckAmount: '$3,261.10',
+      payCheckDt: '03/01/2019',
+      payCheckId: '002',
+      payCheckReturnFiche: 'C',
+      payCheckType: 'Compensation & Pension - Recurring',
     },
     {
-      id: 3,
-      date: 'June 4, 2020',
-      amount: '$4,321',
-      type: 'Componsation & Pension Recurring',
-      method: 'Direct Deposit',
-      bank: 'USAA Federal Savings Bank',
-      account: '1234567890999',
-      rowClass: 'class',
+      payCheckAmount: '$177.60',
+      payCheckDt: '11/09/2018',
+      payCheckId: '003',
+      payCheckReturnFiche: 'C',
+      payCheckType: 'Compensation & Pension - Retroactive',
     },
     {
-      id: 4,
-      date: 'June 3, 2020',
-      amount: '$4,321',
-      type: 'Componsation & Pension Recurring',
-      method: 'Direct Deposit',
-      bank: 'USAA Federal Savings Bank',
-      account: '1234567890999',
-      rowClass: 'class',
+      payCheckAmount: '$3,261.10',
+      payCheckDt: '04/01/2019',
+      payCheckId: '001',
+      payCheckReturnFiche: 'C',
+      payCheckType: 'Compensation & Pension - Recurring',
     },
     {
-      id: 5,
-      date: 'June 4, 2020',
-      amount: '$4,321',
-      type: 'Componsation & Pension Recurring',
-      method: 'Direct Deposit',
-      bank: 'USAA Federal Savings Bank',
-      account: '1234567890999',
-      rowClass: 'class',
+      payCheckAmount: '$3,261.10',
+      payCheckDt: '03/01/2019',
+      payCheckId: '002',
+      payCheckReturnFiche: 'C',
+      payCheckType: 'Compensation & Pension - Recurring',
     },
     {
-      id: 5,
-      date: 'June 3, 2020',
-      amount: '$4,321',
-      type: 'Componsation & Pension Recurring',
-      method: 'Direct Deposit',
-      bank: 'USAA Federal Savings Bank',
-      account: '1234567890999',
-      rowClass: 'class',
+      payCheckAmount: '$177.60',
+      payCheckDt: '11/09/2018',
+      payCheckId: '003',
+      payCheckReturnFiche: 'C',
+      payCheckType: 'Compensation & Pension - Retroactive',
     },
     {
-      id: 6,
-      date: 'June 4, 2020',
-      amount: '$4,321',
-      type: 'Componsation & Pension Recurring',
-      method: 'Direct Deposit',
-      bank: 'USAA Federal Savings Bank',
-      account: '1234567890999',
-      rowClass: 'class',
+      payCheckAmount: '$3,261.10',
+      payCheckDt: '04/01/2019',
+      payCheckId: '001',
+      payCheckReturnFiche: 'C',
+      payCheckType: 'Compensation & Pension - Recurring',
     },
     {
-      id: 7,
-      date: 'June 3, 2020',
-      amount: '$4,321',
-      type: 'Componsation & Pension Recurring',
-      method: 'Direct Deposit',
-      bank: 'USAA Federal Savings Bank',
-      account: '1234567890999',
-      rowClass: 'class',
+      payCheckAmount: '$3,261.10',
+      payCheckDt: '03/01/2019',
+      payCheckId: '002',
+      payCheckReturnFiche: 'C',
+      payCheckType: 'Compensation & Pension - Recurring',
     },
     {
-      id: 8,
-      date: 'June 4, 2020',
-      amount: '$4,321',
-      type: 'Componsation & Pension Recurring',
-      method: 'Direct Deposit',
-      bank: 'USAA Federal Savings Bank',
-      account: '1234567890999',
-      rowClass: 'class',
-    },
-    {
-      id: 9,
-      date: 'June 3, 2020',
-      amount: '$4,321',
-      type: 'Componsation & Pension Recurring',
-      method: 'Direct Deposit',
-      bank: 'USAA Federal Savings Bank',
-      account: '1234567890999',
-      rowClass: 'class',
-    },
-    {
-      id: 10,
-      date: 'June 4, 2020',
-      amount: '$4,321',
-      type: 'Componsation & Pension Recurring',
-      method: 'Direct Deposit',
-      bank: 'USAA Federal Savings Bank',
-      account: '1234567890999',
-      rowClass: 'class',
-    },
-    {
-      id: 11,
-      date: 'June 3, 2020',
-      amount: '$4,321',
-      type: 'Componsation & Pension Recurring',
-      method: 'Direct Deposit',
-      bank: 'USAA Federal Savings Bank',
-      account: '1234567890999',
-      rowClass: 'class',
+      payCheckAmount: '$177.60',
+      payCheckDt: '11/09/2018',
+      payCheckId: '003',
+      payCheckReturnFiche: 'C',
+      payCheckType: 'Compensation & Pension - Retroactive',
     },
   ],
-  returned: [
+  returnPayments: [
     {
-      id: 1,
-      date: 'June 4, 2020',
-      amount: '$4,321',
-      type: 'Componsation & Pension Recurring',
-      method: 'Direct Deposit',
-      bank: 'USAA Federal Savings Bank',
-      account: '1234567890999',
-      rowClass: 'class',
+      returnedCheckAmount: null,
+      returnedCheckCancelDt: '12/01/2012',
+      returnedCheckIssueDt: null,
+      returnedCheckNum: '12365494',
+      returnedCheckRo: '17',
+      returnedCheckReason: '6',
+      returnedCheckReturnFiche: 'B',
+      returnedCheckType: 'Other',
     },
     {
-      id: 2,
-      date: 'June 3, 2020',
-      amount: '$4,321',
-      type: 'Componsation & Pension Recurring',
-      method: 'Direct Deposit',
-      bank: 'USAA Federal Savings Bank',
-      account: '1234567890999',
-      rowClass: 'class',
+      returnedCheckAmount: '$50.00',
+      returnedCheckCancelDt: null,
+      returnedCheckIssueDt: '11/25/2012',
+      returnedCheckNum: '12365494',
+      returnedCheckRo: '17',
+      returnedCheckReason: '6',
+      returnedCheckReturnFiche: 'B',
+      returnedCheckType: 'Post-9/11 Educational Assistance',
     },
   ],
 };
+
+export const fields = [
+  {
+    label: 'Date',
+    value: 'date',
+  },
+  {
+    label: 'Amount',
+    value: 'amount',
+  },
+  {
+    label: 'Type',
+    value: 'type',
+  },
+  {
+    label: 'Method',
+    value: 'method',
+  },
+  {
+    label: 'Bank name',
+    value: 'bank',
+  },
+  {
+    label: 'Account',
+    value: 'account',
+  },
+];
